hsla: Just store floats
We are using floats for rgb, and we don't need more precision
for hsl colors either. We use hsl for computing color expressions
like shade(), lighter() and darker(), which are not precisely
specified anyway.

This commit updates the one test where the output changes a
tiny bit due to this.

+++ gtk/gtkhslaprivate.h
@@ -25,17 +25,17 @@ G_BEGIN_DECLS
 typedef struct _GtkHSLA GtkHSLA;
 
 struct _GtkHSLA {
-  double hue;
-  double saturation;
-  double lightness;
-  double alpha;
+  float hue;
+  float saturation;
+  float lightness;
+  float alpha;
 };
 
 void            _gtk_hsla_init              (GtkHSLA          *hsla,
-                                             double            hue,
-                                             double            saturation,
-                                             double            lightness,
-                                             double            alpha);
+                                             float             hue,
+                                             float             saturation,
+                                             float             lightness,
+                                             float             alpha);
 void            _gtk_hsla_init_from_rgba    (GtkHSLA          *hsla,
                                              const GdkRGBA    *rgba);
 /* Yes, I can name that function like this! */
@@ -44,7 +44,7 @@ void            _gdk_rgba_init_from_hsla    (GdkRGBA          *rgba,
 
 void            _gtk_hsla_shade             (GtkHSLA          *dest,
                                              const GtkHSLA    *src,
-                                             double            factor);
+                                             float             factor);
 
 G_END_DECLS
